requirements.txt
altgraph==0.17.3
appdirs==1.4.4
attrs==22.1.0
auto-py-to-exe==2.24.1
bottle==0.12.23
bottle-websocket==0.2.9
cattrs==22.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
Eel==0.14.0
exceptiongroup==1.0.4
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
idna==3.4
lxml==4.9.1
macholib==1.16.2
numpy==1.23.4
pandas==1.5.1
pydantic==1.10.2
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
pyparsing==3.0.9
PySimpleGUI==4.60.4
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-cache==0.9.7
sdmx1==2.7.0
six==1.16.0
typing_extensions==4.4.0
url-normalize==1.4.3
urllib3==1.26.12
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.5.2
